requirements.txt
requests==2.31.0
flask==3.0.0
anthropic>=0.40.0
telethon>=1.36.0
gspread>=6.0.0
feedparser>=6.0.0
websocket-client>=1.7.0
python-dotenv>=1.0.0
